requirements.txt
openai==0.27.8
python-dotenv==1.0.0
alive-progress==3.1.4
plotly==5.15.0
kaleido==0.2.1